requirements.txt
tweepy
flask
flask-cors
python-dotenv
spacy
vaderSentiment
feedparser
beautifulsoup4
requests
langdetect
deep-translator
transformers
torch
tensorflow
google-api-python-client
youtube-transcript-api
tenacity
nltk
schedule
